\name{MSCMGARCH-package}
\alias{MSCMGARCH-package}
\alias{MSCMGARCH}
\docType{package}
\title{
\packageTitle{MSCMGARCH}
}
\description{
\packageDescription{MSCMGARCH}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MSCMGARCH}
\packageIndices{MSCMGARCH}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MSCMGARCH}

Maintainer: \packageMaintainer{MSCMGARCH}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
